refactor transformer and add some docs
