+++ transformer/src/transformers/file2ByteArray.ts
